# backend/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
...
		"/problem/{id}/user-info": {
			"get": {
				"operationId": "GetUserInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"number_tries": {
											"type": "number",
											"format": "double"
										},
										"status": {
											"type": "string"
										}
									},
									"required": [
										"number_tries",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/problem/{id}/samples": {
			"get": {
				"operationId": "GetSamples",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"problems": {
											"items": {},
											"type": "array"
										}
									},
									"required": [
										"problems"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/problem/{id}/skeleton/{language}": {
			"get": {
				"operationId": "GetSkeletonCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"text/plain": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "language",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/problem/{id}/submits": {
			"get": {
				"operationId": "GetSubmits",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "string"
										},
										"submits": {
											"items": {},
											"type": "array"
										}
									},
									"required": [
										"status",
										"submits"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}